+++ pyproject.toml
@@ -71,7 +71,7 @@ llama-index = "^0.11.9"
 llama-index-embeddings-openai = "^0.2.5"
 llama-index-embeddings-ollama = "^0.3.1"
 #llama-index-embeddings-huggingface = {version = "^0.2.0", optional = true}
-wikipedia = "^1.4.0"
+wikipedia = {version = "^1.4.0", optional = true}
 
 [tool.poetry.extras]
 #local = ["llama-index-embeddings-huggingface"]
@@ -83,6 +83,7 @@ autogen = ["pyautogen"]
 qdrant = ["qdrant-client"]
 ollama = ["llama-index-embeddings-ollama"]
 crewai-tools = ["crewai", "docker", "crewai-tools"]
+tests = ["wikipedia"]
 
 [tool.poetry.group.dev.dependencies]
 black = "^24.4.2"
